requirements.txt
torch
torchvision
gymnasium[box2d]
numpy
matplotlib
hydra-core
tqdm